Move `runtime/opsetx/ops.py` and update inits to import from new location
