+++ src/bindings/python/src/openvino/opset10/__init__.py
