+++ src/bindings/python/src/openvino/opset10/ops.py
