+++ src/bindings/python/src/openvino/opset12/__init__.py
