+++ src/bindings/python/src/openvino/opset12/ops.py
