+++ src/bindings/python/src/openvino/opset13/ops.py
@@ -12,8 +12,8 @@ import numpy as np
 log = logging.getLogger(__name__)
 
 from openvino.runtime import Node, Shape, Type, Output, Tensor
-from openvino.runtime.op import Constant, Result
-from openvino.runtime.opset1 import convert_like
+from openvino.op import Constant, Result
+from openvino.opset1 import convert_like
 from openvino.runtime.opset_utils import _get_node_factory
 from openvino.runtime.utils.decorators import binary_op, nameable_op, unary_op, overloading
 from openvino.runtime.utils.types import (
