+++ src/bindings/python/src/openvino/opset15/__init__.py
