+++ src/bindings/python/src/openvino/opset16/__init__.py
