+++ src/bindings/python/src/openvino/opset3/__init__.py
