+++ src/bindings/python/src/openvino/opset3/ops.py
@@ -9,7 +9,7 @@ import numpy as np
 from functools import partial
 
 from openvino.runtime import Node, Shape
-from openvino.runtime.op import Constant, Parameter
+from openvino.op import Constant, Parameter
 from openvino.runtime.opset_utils import _get_node_factory
 from openvino.runtime.utils.decorators import binary_op, nameable_op, unary_op
 from openvino.runtime.utils.input_validation import (
