+++ src/bindings/python/src/openvino/opset4/__init__.py
