+++ src/bindings/python/src/openvino/opset5/__init__.py
