+++ src/bindings/python/src/openvino/opset6/__init__.py
