+++ src/bindings/python/src/openvino/opset6/ops.py
@@ -10,9 +10,9 @@ from typing import Optional, Union
 from functools import partial, singledispatch
 
 from openvino.runtime import Node, Type, PartialShape, Output, Shape
-from openvino.runtime.op import assign, Constant, Parameter
-from openvino.runtime.op import read_value as _read_value
-from openvino.runtime.op.util import VariableInfo, Variable
+from openvino.op import assign, Constant, Parameter
+from openvino.op import read_value as _read_value
+from openvino.op.util import VariableInfo, Variable
 from openvino.runtime.opset_utils import _get_node_factory
 from openvino.runtime.utils.decorators import nameable_op, overloading
 from openvino.runtime.utils.types import (
