+++ src/bindings/python/src/openvino/opset7/__init__.py
