+++ src/bindings/python/src/openvino/opset7/ops.py
@@ -8,7 +8,7 @@ from typing import Callable, Iterable, List, Optional, Set, Union
 
 import numpy as np
 from openvino.runtime import Node, Shape
-from openvino.runtime.op import Constant, Parameter
+from openvino.op import Constant, Parameter
 from openvino.runtime.opset_utils import _get_node_factory
 from openvino.runtime.utils.decorators import binary_op, nameable_op, unary_op
 from openvino.runtime.utils.input_validation import (
