+++ src/bindings/python/src/openvino/opset8/ops.py
@@ -8,7 +8,7 @@ from typing import List, Optional, Tuple
 
 import numpy as np
 from openvino.runtime.exceptions import UserInputError
-from openvino.runtime.op import Constant, Parameter, if_op
+from openvino.op import Constant, Parameter, if_op
 from openvino.runtime import Node
 from openvino.runtime.opset_utils import _get_node_factory
 from openvino.runtime.utils.decorators import nameable_op
